+++ include/codi/tools/helpers/externalFunctionHelper.hpp
@@ -75,6 +75,8 @@ namespace codi {
    *
    * The ExternalFunctionHelper works with all tapes. It is also able to handle situations where the tape is currently
    * not recording. All necessary operations are performed in such a case but no external function is recorded.
+   * If enablePrimalValueTapeSpecialization is active, primal values are recovered from the tape instead of being stored 
+   * in the ExternalFunctionHelper instance, reducing its memory footprint.
    *
    * The storing of primal inputs and outputs can be disabled. Outputs can be discarded if they are recomputed in the
    * derivative computation or if the derivative does not depend on them. Inputs can be discarded if the derivative does
@@ -106,6 +108,9 @@ namespace codi {
       using PrimalFunc = void (*)(Real const* x, size_t m, Real* y, size_t n, ExternalFunctionUserData* d);
 
     private:
+
+      static constexpr bool IsPrimalValueTape = TapeTraits::IsPrimalValueTape<Tape>::value;
+
       struct EvalData {
         public:
 
@@ -122,6 +127,10 @@ namespace codi {
 
           ExternalFunctionUserData userData;
 
+          bool provideInputValues;
+          bool provideOutputValues;
+          bool getPrimalsFromPrimalValueVector;
+
           EvalData()
               : inputIndices(0),
                 outputIndices(0),
@@ -130,7 +139,10 @@ namespace codi {
                 oldPrimals(0),
                 reverseFunc(nullptr),
                 forwardFunc(nullptr),
-                primalFunc(nullptr) {}
+                primalFunc(nullptr),
+                provideInputValues(true),
+                provideOutputValues(true),
+                getPrimalsFromPrimalValueVector(false) {}
 
           static void delFunc(Tape* t, void* d) {
             CODI_UNUSED(t);
@@ -151,25 +163,21 @@ namespace codi {
             }
           }
 
-          void evalForwFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
+          CODI_INLINE void evalForwFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
             CODI_UNUSED(t);
 
             std::vector<Real> x_d(inputIndices.size());
             std::vector<Real> y_d(outputIndices.size());
 
-            if (TapeTraits::IsPrimalValueTape<Tape>::value) {
-              for (size_t i = 0; i < inputIndices.size(); ++i) {
-                inputValues[i] = ra->getPrimal(inputIndices[i]);
-              }
-            }
+            initRun(ra);
 
             for (size_t dim = 0; dim < ra->getVectorSize(); ++dim) {
               for (size_t i = 0; i < inputIndices.size(); ++i) {
                 x_d[i] = ra->getAdjoint(inputIndices[i], dim);
               }
 
-              forwardFunc(inputValues.data(), x_d.data(), inputIndices.size(), outputValues.data(), y_d.data(), outputIndices.size(),
-                          &userData);
+              forwardFunc(inputValues.data(), x_d.data(), inputIndices.size(), outputValues.data(), y_d.data(),
+                          outputIndices.size(), &userData);
 
               for (size_t i = 0; i < outputIndices.size(); ++i) {
                 ra->resetAdjoint(outputIndices[i], dim);
@@ -177,11 +185,7 @@ namespace codi {
               }
             }
 
-            if (TapeTraits::IsPrimalValueTape<Tape>::value) {
-              for (size_t i = 0; i < outputIndices.size(); ++i) {
-                ra->setPrimal(outputIndices[i], outputValues[i]);
-              }
-            }
+            finalizeRun(ra);
           }
 
           static void evalPrimFuncStatic(Tape* t, void* d, VectorAccessInterface<Real, Identifier>* ra) {
@@ -195,22 +199,14 @@ namespace codi {
             }
           }
 
-          void evalPrimFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
+          CODI_INLINE void evalPrimFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
             CODI_UNUSED(t);
 
-            if (TapeTraits::IsPrimalValueTape<Tape>::value) {
-              for (size_t i = 0; i < inputIndices.size(); ++i) {
-                inputValues[i] = ra->getPrimal(inputIndices[i]);
-              }
-            }
+            initRun(ra);
 
             primalFunc(inputValues.data(), inputIndices.size(), outputValues.data(), outputIndices.size(), &userData);
 
-            if (TapeTraits::IsPrimalValueTape<Tape>::value) {
-              for (size_t i = 0; i < outputIndices.size(); ++i) {
-                ra->setPrimal(outputIndices[i], outputValues[i]);
-              }
-            }
+            finalizeRun(ra);
           }
 
           static void evalRevFuncStatic(Tape* t, void* d, VectorAccessInterface<Real, Identifier>* ra) {
@@ -224,11 +220,13 @@ namespace codi {
             }
           }
 
-          void evalRevFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
+          CODI_INLINE void evalRevFunc(Tape* t, VectorAccessInterface<Real, Identifier>* ra) {
             CODI_UNUSED(t);
 
             std::vector<Real> x_b(inputIndices.size());
             std::vector<Real> y_b(outputIndices.size());
+
+            initRun(ra, true);
 
             for (size_t dim = 0; dim < ra->getVectorSize(); ++dim) {
               for (size_t i = 0; i < outputIndices.size(); ++i) {
@@ -236,18 +234,63 @@ namespace codi {
                 ra->resetAdjoint(outputIndices[i], dim);
               }
 
-              reverseFunc(inputValues.data(), x_b.data(), inputIndices.size(), outputValues.data(), y_b.data(), outputIndices.size(),
-                          &userData);
+              reverseFunc(inputValues.data(), x_b.data(), inputIndices.size(), outputValues.data(), y_b.data(),
+                          outputIndices.size(), &userData);
 
               for (size_t i = 0; i < inputIndices.size(); ++i) {
                 ra->updateAdjoint(inputIndices[i], dim, x_b[i]);
               }
             }
 
-            if (Tape::RequiresPrimalRestore) {
+            finalizeRun(ra, true);
+          }
+
+        private:
+
+          CODI_INLINE void initRun(VectorAccessInterface<Real, Identifier>* ra, bool isReverse = false) {
+            if (getPrimalsFromPrimalValueVector && provideOutputValues) {
+              outputValues.resize(outputIndices.size());
+
+              if (isReverse) {  // Provide result values for reverse evaluations.
+                for (size_t i = 0; i < outputIndices.size(); ++i) {
+                  outputValues[i] = ra->getPrimal(outputIndices[i]);
+                }
+              }
+            }
+
+            // Restore the old primals for reverse evaluations, before the inputs are read.
+            if (isReverse && Tape::RequiresPrimalRestore) {
               for (size_t i = 0; i < outputIndices.size(); ++i) {
                 ra->setPrimal(outputIndices[i], oldPrimals[i]);
               }
+            }
+
+            if (getPrimalsFromPrimalValueVector && provideInputValues) {
+              inputValues.resize(inputIndices.size());
+
+              for (size_t i = 0; i < inputIndices.size(); ++i) {
+                inputValues[i] = ra->getPrimal(inputIndices[i]);
+              }
+            }
+          }
+
+          CODI_INLINE void finalizeRun(VectorAccessInterface<Real, Identifier>* ra, bool isReverse = false) {
+            if (getPrimalsFromPrimalValueVector && !isReverse) {
+              for (size_t i = 0; i < outputIndices.size(); ++i) {
+                if (Tape::RequiresPrimalRestore) {
+                  oldPrimals[i] = ra->getPrimal(outputIndices[i]);
+                }
+                ra->setPrimal(outputIndices[i], outputValues[i]);
+              }
+            }
+
+            if (getPrimalsFromPrimalValueVector && provideInputValues) {
+              inputValues.clear();
+              inputValues.shrink_to_fit();
+            }
+            if (getPrimalsFromPrimalValueVector && provideOutputValues) {
+              outputValues.clear();
+              outputValues.shrink_to_fit();
             }
           }
       };
@@ -256,9 +299,9 @@ namespace codi {
 
       std::vector<Type*> outputValues;  ///< References to output values.
 
-      bool storeInputPrimals;     ///< If input primals are stored. Can be disabled by the user.
-      bool storeOutputPrimals;    ///< If output primals are stored. Can be disabled by the user.
-      bool primalFuncUsesADType;  ///< If a primal call with a self-implemented function will be done.
+      bool storeInputPrimals;              ///< If input primals are stored. Can be disabled by the user.
+      bool storeOutputPrimals;             ///< If output primals are stored. Can be disabled by the user.
+      bool storeInputOutputForPrimalEval;  ///< If a primal call with a self-implemented function will be done.
 
       EvalData* data;  ///< External function data.
 
@@ -269,7 +312,7 @@ namespace codi {
           : outputValues(),
             storeInputPrimals(true),
             storeOutputPrimals(true),
-            primalFuncUsesADType(primalFuncUsesADType),
+            storeInputOutputForPrimalEval(!primalFuncUsesADType),
             data(nullptr) {
         data = new EvalData();
       }
@@ -279,32 +322,55 @@ namespace codi {
         delete data;
       }
 
+      /// Load primal values from the primal value vector of the tape. (Has no effect on Jacobian tapes.)
+      void enablePrimalValueTapeSpecialization() {
+        if (IsPrimalValueTape) {
+          storeInputPrimals = false;
+          storeOutputPrimals = false;
+          data->getPrimalsFromPrimalValueVector = true;
+        }
+      }
+
       /// Do not store primal input values. In function calls, pointers to primal inputs will be null.
       void disableInputPrimalStore() {
         storeInputPrimals = false;
+        data->provideInputValues = false;
       }
 
       /// Do not store primal output values. In function calls, pointers to primal outputs will be null.
       void disableOutputPrimalStore() {
         storeOutputPrimals = false;
+        data->provideOutputValues = false;
       }
 
       /// Add an input value.
-      void addInput(Type const& input) {
+      CODI_INLINE void addInput(Type const& input) {
         if (Type::getTape().isActive()) {
-          data->inputIndices.push_back(input.getIdentifier());
+          Identifier identifier = input.getIdentifier();
+          if (!Type::getTape().isIdentifierActive(identifier)) {
+            // Register input values for primal value tapes when they are restored from the tape, otherwise the primal
+            // values can not be restored. For a lot of inactive inputs, this can inflate the number of identifiers
+            // quite a lot. This is especially true for reuse index tapes.
+            if (data->getPrimalsFromPrimalValueVector) {
+              Type temp = input;
+              Type::getTape().registerInput(temp);
+              identifier = temp.getIdentifier();
+            }
+          }
+
+          data->inputIndices.push_back(identifier);
         }
 
         // Ignore the setting at this place and the active check,
         // we might need the values for the evaluation.
-        if (!primalFuncUsesADType || storeInputPrimals) {
+        if (storeInputOutputForPrimalEval || storeInputPrimals) {
           data->inputValues.push_back(input.getValue());
         }
       }
 
     private:
 
-      void addOutputToData(Type& output) {
+      CODI_INLINE void addOutputToData(Type& output) {
         Real oldPrimal = Type::getTape().registerExternalFunctionOutput(output);
 
         data->outputIndices.push_back(output.getIdentifier());
@@ -319,15 +385,15 @@ namespace codi {
     public:
 
       /// Add an output value.
-      void addOutput(Type& output) {
-        if (Type::getTape().isActive()) {
+      CODI_INLINE void addOutput(Type& output) {
+        if (Type::getTape().isActive() || storeInputOutputForPrimalEval) {
           outputValues.push_back(&output);
         }
       }
 
       /// Add user data. See ExternalFunctionUserData for details.
       template<typename Data>
-      void addUserData(Data const& data) {
+      CODI_INLINE void addUserData(Data const& data) {
         this->data->userData.addData(data);
       }
 
@@ -340,7 +406,7 @@ namespace codi {
       /// This is intended for primal functions that are implemented with the AD type. It is ensured that no data is
       /// recorded on the tape. All output values are registered as outputs of this external function.
       template<typename FuncObj, typename... Args>
-      void callPrimalFuncWithADType(FuncObj& func, Args&&... args) {
+      CODI_INLINE void callPrimalFuncWithADType(FuncObj& func, Args&&... args) {
         bool isTapeActive = Type::getTape().isActive();
 
         if (isTapeActive) {
@@ -360,8 +426,8 @@ namespace codi {
 
       /// Call the primal function with the values extracted from the inputs. The output values are set on the specified
       /// outputs and registered as outputs of this external functions.
-      void callPrimalFunc(PrimalFunc func) {
-        if (!primalFuncUsesADType) {
+      CODI_INLINE void callPrimalFunc(PrimalFunc func) {
+        if (storeInputOutputForPrimalEval) {
           // Store the primal function in the external function data so that it can be used for primal evaluations of
           // the tape.
           data->primalFunc = func;
@@ -386,7 +452,8 @@ namespace codi {
       }
 
       /// Add the external function to the tape.
-      void addToTape(ReverseFunc reverseFunc, ForwardFunc forwardFunc = nullptr, PrimalFunc primalFunc = nullptr) {
+      CODI_INLINE void addToTape(ReverseFunc reverseFunc, ForwardFunc forwardFunc = nullptr,
+                                 PrimalFunc primalFunc = nullptr) {
         if (Type::getTape().isActive()) {
           data->reverseFunc = reverseFunc;
           data->forwardFunc = forwardFunc;
@@ -400,6 +467,7 @@ namespace codi {
           // Clear the primal values if they are not required.
           if (!storeInputPrimals) {
             data->inputValues.clear();
+            data->inputValues.shrink_to_fit();
           }
 
           Type::getTape().pushExternalFunction(